# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Centinela",
    "description": "Personal service-monitoring platform with health checks and history.",
    "version": "0.6.0"
  },
  "paths": {
...
    "/api/v1/ai/config": {
      "get": {
        "tags": [
          "ai-settings"
        ],
        "summary": "Get Ai Config",
        "operationId": "get_ai_config_api_v1_ai_config_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AISettingsRead"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "ai-settings"
        ],
        "summary": "Update Ai Config",
        "operationId": "update_ai_config_api_v1_ai_config_put",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AISettingsUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AISettingsRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ai/config/credential": {
      "delete": {
        "tags": [
          "ai-settings"
        ],
        "summary": "Delete Ai Credential",
        "operationId": "delete_ai_credential_api_v1_ai_config_credential_delete",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AISettingsRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ai/config/test": {
      "post": {
        "tags": [
          "ai-settings"
        ],
        "summary": "Test Ai Config",
        "operationId": "test_ai_config_api_v1_ai_config_test_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIProviderTestRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AIProviderTestRead": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "provider": {
            "$ref": "#/components/schemas/ProviderType"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "latency_ms": {
            "type": "integer",
            "title": "Latency Ms"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "provider",
          "model",
          "latency_ms"
        ],
        "title": "AIProviderTestRead"
      },
      "AISettingsRead": {
        "properties": {
          "provider": {
            "$ref": "#/components/schemas/ProviderType"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "summary_language": {
            "$ref": "#/components/schemas/SummaryLanguage"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "credential_required": {
            "type": "boolean",
            "title": "Credential Required"
          },
          "credential_configured": {
            "type": "boolean",
            "title": "Credential Configured"
          },
          "api_key_hint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key Hint"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "provider",
          "model",
          "summary_language",
          "enabled",
          "credential_required",
          "credential_configured",
          "api_key_hint",
          "updated_at"
        ],
        "title": "AISettingsRead"
      },
      "AISettingsUpdate": {
        "properties": {
          "provider": {
            "$ref": "#/components/schemas/ProviderType"
          },
          "model": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Model"
          },
          "summary_language": {
            "$ref": "#/components/schemas/SummaryLanguage",
            "default": "es"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          },
          "api_key": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000,
                "minLength": 8
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key"
          }
        },
        "type": "object",
        "required": [
          "provider",
          "model"
        ],
        "title": "AISettingsUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProviderType": {
        "type": "string",
        "enum": [
          "ollama",
          "openai",
          "anthropic"
        ],
        "title": "ProviderType",
        "description": "AI providers supported by the first multi-provider release."
      },
...
      "SummaryLanguage": {
        "type": "string",
        "enum": [
          "es",
          "en"
        ],
        "title": "SummaryLanguage",
        "description": "Languages supported by the incident prompt and UI."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}